file refactor and reorganizing
