+++ app/admin/aboutUs/AboutUs.tsx
@@ -1,41 +1,41 @@
 "use client";
 
-import { useOurWorkLogic } from "./OurWorkLogic"; // Import the logic file
-import { useCollapse } from "../supportFunctions/ToggleCollase";
-import blankImage from "../global-images/blank-person.png";
+import { useAboutUsLogic } from "./AboutUsLogic"; // Import the logic file
+import { useCollapse } from "../../supportFunctions/ToggleCollase";
+import blankImage from "../../global-images/blank-person.png";
 import "@aws-amplify/ui-react/styles.css";
-import "../page.module.css";
-import "./admin.css";
+import "../../page.module.css";
+import "../admin.css";
 
-export default function OurWork() {
+export default function AboutUs() {
   const {
-    ourWorks,
-    editingOurWorks,
+    emps,
+    editingEmps,
     picture,
+    name,
+    title,
     description,
-    business,
     setPicture,
+    setName,
+    setTitle,
     setDescription,
-    setBusiness,
-    handleEditChangeOurWork,
-    handleSaveChangesOurWork,
-    handleEditToggleOurWork,
-    handleCancelEditOurWork,
-    handleDeleteOurWork,
-    handleOurWorkSubmit,
-  } = useOurWorkLogic();
+    handleAboutUsSubmit,
+    handleEditChangeEmp,
+    handleSaveChangesEmp,
+    handleEditToggleEmp,
+    handleCancelEditEmp,
+    deleteEntry,
+  } = useAboutUsLogic();
 
-  const { isContentCollapsed, toggleCollapse } = useCollapse();
+  const { isContentCollapsed, toggleCollapse} = useCollapse();
 
   return (
     <>
       <div className="div">
         <h2 className="admin-h2" onClick={toggleCollapse}>
-          Our Work{" "}
+          About Us{" "}
           <span
-            className={`dropdown-arrow ${
-              isContentCollapsed ? "collapsed" : ""
-            }`}
+            className={`dropdown-arrow ${isContentCollapsed ? ".collapsed" : ""}`}
             style={{
               display: "inline-block",
               marginLeft: "8px",
@@ -46,36 +46,34 @@ export default function OurWork() {
             ▼
           </span>
         </h2>
+
         {/* Collapsible Section */}
         {isContentCollapsed && (
           <>
-            <h3 className="admin-h3">Current Business / Organizations</h3>
+            <br />
+            <h3 className="admin-h3">Team Members</h3>
             <table className="admin-about-us-table">
               <thead>
                 <tr>
-                  <th>Select</th>
-                  <th>Business</th>
+                  <th>Picture</th>
+                  <th>Name</th>
+                  <th>Title</th>
                   <th>Description</th>
                   <th>Actions</th>
                 </tr>
               </thead>
               <tbody>
-                {ourWorks.map((ourWork) => {
-                  const editingourWork =
-                    editingOurWorks.get(ourWork.id) || ourWork;
+                {emps.map((emp) => {
+                  const editingEmp = editingEmps.get(emp.id) || emp;
                   return (
-                    <tr key={ourWork.id}>
+                    <tr key={emp.id}>
                       <td>
-                        {editingOurWorks.has(ourWork.id) ? (
+                        {editingEmps.has(emp.id) ? (
                           <input
                             type="text"
-                            value={editingourWork.picture || ""}
+                            value={editingEmp.picture || ""}
                             onChange={(e) =>
-                              handleEditChangeOurWork(
-                                ourWork.id,
-                                "picture",
-                                e.target.value
-                              )
+                              handleEditChangeEmp(emp.id, "picture", e.target.value)
                             }
                           />
                         ) : (
@@ -87,69 +85,60 @@ export default function OurWork() {
                         )}
                       </td>
                       <td>
-                        {editingOurWorks.has(ourWork.id) ? (
+                        {editingEmps.has(emp.id) ? (
                           <input
                             type="text"
-                            value={editingourWork.business || ""}
+                            value={editingEmp.name || ""}
                             onChange={(e) =>
-                              handleEditChangeOurWork(
-                                ourWork.id,
-                                "business",
-                                e.target.value
-                              )
+                              handleEditChangeEmp(emp.id, "name", e.target.value)
                             }
                           />
                         ) : (
-                          ourWork.business
+                          emp.name
                         )}
                       </td>
                       <td>
-                        {editingOurWorks.has(ourWork.id) ? (
+                        {editingEmps.has(emp.id) ? (
                           <input
                             type="text"
-                            value={editingourWork.description || ""}
+                            value={editingEmp.title || ""}
                             onChange={(e) =>
-                              handleEditChangeOurWork(
-                                ourWork.id,
-                                "description",
-                                e.target.value
-                              )
+                              handleEditChangeEmp(emp.id, "title", e.target.value)
                             }
                           />
                         ) : (
-                          ourWork.description
+                          emp.title
                         )}
                       </td>
                       <td>
-                        {editingOurWorks.has(ourWork.id) ? (
+                        {editingEmps.has(emp.id) ? (
+                          <input
+                            type="text"
+                            value={editingEmp.description || ""}
+                            onChange={(e) =>
+                              handleEditChangeEmp(emp.id, "description", e.target.value)
+                            }
+                          />
+                        ) : (
+                          emp.description
+                        )}
+                      </td>
+                      <td>
+                        {editingEmps.has(emp.id) ? (
                           <>
-                            <button
-                              onClick={() =>
-                                handleSaveChangesOurWork(ourWork.id)
-                              }
-                            >
+                            <button onClick={() => handleSaveChangesEmp(emp.id)}>
                               Save Changes
                             </button>
-                            <button
-                              onClick={() =>
-                                handleCancelEditOurWork(ourWork.id)
-                              }
-                            >
+                            <button onClick={() => handleCancelEditEmp(emp.id)}>
                               Cancel
                             </button>
                           </>
                         ) : (
                           <>
-                            <button
-                              onClick={() =>
-                                handleEditToggleOurWork(ourWork.id)
-                              }
-                            >
+                            <button onClick={() => handleEditToggleEmp(emp.id)}>
                               Edit
                             </button>
-                            <button
-                              onClick={() => handleDeleteOurWork(ourWork.id)}
-                            >
+                            <button onClick={() => deleteEntry(emp.id)}>
                               Delete
                             </button>
                           </>
@@ -161,32 +150,43 @@ export default function OurWork() {
               </tbody>
             </table>
             <br />
-            <form onSubmit={handleOurWorkSubmit} className="about-us-form">
-              <h3 className="admin-h3">Add Business</h3>
+            <form onSubmit={handleAboutUsSubmit} className="about-us-form">
+              <h3 className="admin-h3">Add Team Member</h3>
               <div className="form-group">
-                <label htmlFor="pictures">Picture URL:</label>
+                <label htmlFor="picture">Picture URL:</label>
                 <input
                   id="picture"
                   type="text"
                   value={picture}
                   onChange={(e) => setPicture(e.target.value)}
-                  placeholder="Enter Picture URL"
+                  placeholder="Enter picture URL"
                   className="form-input"
                 />
               </div>
               <div className="form-group">
-                <label htmlFor="business">Business:</label>
+                <label htmlFor="name">Name:</label>
                 <input
-                  id="business"
+                  id="name"
                   type="text"
-                  value={business}
-                  onChange={(e) => setBusiness(e.target.value)}
-                  placeholder="Enter in Business name"
+                  value={name}
+                  onChange={(e) => setName(e.target.value)}
+                  placeholder="Enter name"
                   className="form-input"
                 />
               </div>
               <div className="form-group">
-                <label htmlFor="description">Description:</label>
+                <label htmlFor="title">Title:</label>
+                <input
+                  id="title"
+                  type="text"
+                  value={title}
+                  onChange={(e) => setTitle(e.target.value)}
+                  placeholder="Enter title"
+                  className="form-input"
+                />
+              </div>
+              <div className="form-group">
+                <label htmlFor="description">Description (Optional):</label>
                 <input
                   id="description"
                   type="text"
