+++ app/admin/emailList/EmailList.tsx
@@ -1,10 +1,10 @@
 "use client";
 
 import { useEmailListLogic } from "./EmailListLogic"; // Import the logic file
-import { useCollapse } from "../supportFunctions/ToggleCollase";
+import { useCollapse } from "../../supportFunctions/ToggleCollase";
 import "@aws-amplify/ui-react/styles.css";
-import "../page.module.css";
-import "./admin.css";
+import "../../page.module.css";
+import "../admin.css";
 
 export default function EmailList() {
   const {
