+++ app/admin/emailList/EmailListLogic.ts
